Move documentation to the header files in include/CGAL/
