--- Solver_interface/doc/Solver_interface/CGAL/Diagonalize_traits.h
@@ -1,23 +1,0 @@
-namespace CGAL {
-
-/*!
-\ingroup PkgSolver
-
-The class `Diagonalize_traits` provides an internal
-implementation for the diagonalization of Variance-Covariance
-Matrices.
-
-\cgalModels `DiagonalizeTraits`
-
-*/
-
-
-  
-template <typename FT, unsigned int dim = 3>
-class Diagonalize_traits{
-
-
-};
-
-} // namespace CGAL
-
