--- Solver_interface/doc/Solver_interface/CGAL/Eigen_diagonalize_traits.h
@@ -1,31 +1,0 @@
-namespace CGAL {
-
-/*!
-\ingroup PkgSolver
-
-The class `Eigen_diagonalize_traits` provides an interface to the
-diagonalization of covariance matrices of \ref thirdpartyEigen
-"Eigen".
-
-The version 3.1 (or greater) of \ref thirdpartyEigen "Eigen" must be
-available on the system.
-
-\cgalModels `DiagonalizeTraits`
-
-
-
-\sa http://eigen.tuxfamily.org
-
-*/
-
-
-  
-template <typename FT, unsigned int dim = 3>
-class Eigen_diagonalize_traits{
-
-
-
-};
-
-} // namespace CGAL
-
