+++ Solver_interface/include/CGAL/Diagonalize_traits.h
@@ -27,11 +27,18 @@
 #include <CGAL/double.h>
 
 namespace CGAL {
+/*!
+\ingroup PkgSolver
 
-  /// A model of the concept `DiagonalizeTraits` 
-  /// \cgalModels `DiagonalizeTraits`
+The class `Diagonalize_traits` provides an internal
+implementation for the diagonalization of Variance-Covariance
+Matrices.
 
-template <typename FT, unsigned int dim = 3>
+\cgalModels `DiagonalizeTraits`
+
+*/
+
+  template <typename FT, unsigned int dim = 3>
 class Diagonalize_traits{
 
 public:
