+++ Solver_interface/include/CGAL/Eigen_diagonalize_traits.h
@@ -39,8 +39,21 @@
 
 namespace CGAL {
 
-/// A model of the concept `DiagonalizeTraits` using \ref thirdpartyEigen.
-/// \cgalModels `DiagonalizeTraits`
+/*!
+\ingroup PkgSolver
+
+The class `Eigen_diagonalize_traits` provides an interface to the
+diagonalization of covariance matrices of \ref thirdpartyEigen
+"Eigen".
+
+The version 3.1 (or greater) of \ref thirdpartyEigen "Eigen" must be
+available on the system.
+
+\cgalModels `DiagonalizeTraits`
+
+\sa http://eigen.tuxfamily.org
+
+*/
 
 template <typename FT, unsigned int dim = 3>
 class Eigen_diagonalize_traits{
