+++ Solver_interface/include/CGAL/Eigen_svd.h
@@ -35,6 +35,17 @@
 
 namespace CGAL {
 
+  /*!
+\ingroup PkgSolver
+
+The class `Eigen_svd` provides an algorithm to solve in the least 
+square sense a linear system with a singular value decomposition using 
+\ref thirdpartyEigen. 
+
+\cgalModels `SvdTraits`
+
+*/
+ 
 class Eigen_svd{
 public:
   typedef double FT;
